requirements.txt
ET
icecream
python-dotenv
pathlib
openai
tiktoken
langchain-text-splitters